Pull data from the Environment Agency's API

The loadAPI function is used to interogate the EAs API. Data for all station can be pulled, or when specified to a single site various datasets can be pulled.

loadAPI(
  ID = NULL,
  measure = NULL,
  period = NULL,
  type = NULL,
  datapoints = "standard",
  from = NULL,
  to = NULL,
  lat = NULL,
  long = NULL,
  easting = NULL,
  northing = NULL,
  dist = NULL,
  obsProperty = NULL
)

Arguments

ID

Use to specify a particular WISKI ID or for when downloading all stations ' wiski' or 'nrfa' to filter results

measure

Use this when exporting observations to select the available parameter. Generally 'flow', 'level', or 'groundwater'

period

This is so you can select the time steps available, generally 900 (15 min) or 86400 (daily)

type

Selects the data type, instantaneous, min, max etc.

datapoints

Allows you to decide where in the time series you are pulling the data from

from

First time step of choice

to

Last time step of choice

lat

Latitude coordinate for when searching by geographic area

long

Longitude coordinate for when searching by geographic area

easting

Easting coordinate for when searching by geographic area

northing

Northing coordinate for when searching by geographic area

dist

Distance in km for how far you wish to search

obsProperty

Used to filter the stations when identifying sites. Available metrics; 'waterFlow', 'waterLevel', 'rainfall', 'groundwaterLevel', 'ammonium', 'dissolved-oxygen', 'conductivity', 'ph', 'temperature', 'turbidity', 'nitrate', 'chlorophyll', 'salinity', 'bga', 'fdom'

Value

Examples

loadAPI()
#>        wiskiID                             label      riverName
#>     1: 037048U                   Ulting Sarasota  River Chelmer
#>     2:  510810                    Beggearn Huish Washford River
#>     3:  510810                    Beggearn Huish Washford River
#>     4:   F0803                            Adwick   River Dearne
#>     5:   F0803                            Adwick   River Dearne
#>    ---                                                         
#> 16327:    <NA> SEGGIMOOR BECK_US CAENBY_E_202304 SEGGIMOOR BECK
#> 16328:    <NA> SEGGIMOOR BECK_US CAENBY_E_202304 SEGGIMOOR BECK
#> 16329:    <NA> SEGGIMOOR BECK_US CAENBY_E_202304 SEGGIMOOR BECK
#> 16330:    <NA> SEGGIMOOR BECK_US CAENBY_E_202304 SEGGIMOOR BECK
#> 16331:    <NA> SEGGIMOOR BECK_US CAENBY_E_202304 SEGGIMOOR BECK
#>        observedProperty easting northing      lat      long dateOpened
#>     1:        waterFlow  581271   208591 51.74668  0.624437 2008-10-31
#>     2:       waterLevel  304007   139460 51.14632 -3.373695 1967-01-01
#>     3:        waterFlow  304007   139460 51.14632 -3.373695 1967-01-01
#>     4:       waterLevel  447677   402022 53.51271 -1.282505 1956-10-29
#>     5:        waterFlow  447677   402022 53.51271 -1.282505 1956-10-29
#>    ---                                                                
#> 16327: dissolved-oxygen  499334   389439 53.39244 -0.507715 2023-05-02
#> 16328:     conductivity  499334   389439 53.39244 -0.507715 2023-05-02
#> 16329:               ph  499334   389439 53.39244 -0.507715 2023-05-02
#> 16330:      temperature  499334   389439 53.39244 -0.507715 2023-05-02
#> 16331:        turbidity  499334   389439 53.39244 -0.507715 2023-05-02
#>        catchmentArea nrfaStationID
#>     1:            NA          <NA>
#>     2:          36.3         51003
#>     3:          36.3         51003
#>     4:         311.0         27030
#>     5:         311.0         27030
#>    ---                            
#> 16327:            NA          <NA>
#> 16328:            NA          <NA>
#> 16329:            NA          <NA>
#> 16330:            NA          <NA>
#> 16331:            NA          <NA>

loadAPI(easting = 378235 , northing = 276165, dist = 30, ID = 'nrfa')
#>    wiskiID        label      riverName observedProperty easting northing
#> 1:    2001      Bewdley   River Severn       waterLevel  378235   276165
#> 2:    2001      Bewdley   River Severn        waterFlow  378235   276165
#> 3:    2034  Oak Cottage   Dowles Brook       waterLevel  377105   276425
#> 4:    2034  Oak Cottage   Dowles Brook        waterFlow  377105   276425
#> 5:    2063    Prestwood    River Stour        waterFlow  386705   286155
#> 6:    2193 Wards Bridge   Hadley Brook        waterFlow  387025   263145
#> 7:    2011 Harford Hill River Salwarpe       waterLevel  386865   261935
#> 8:    2011 Harford Hill River Salwarpe        waterFlow  386865   261935
#>         lat      long dateOpened catchmentArea nrfaStationID
#> 1: 52.38307 -2.321186 1921-04-01        4325.0         54001
#> 2: 52.38307 -2.321186 1921-04-01        4325.0         54001
#> 3: 52.38536 -2.337805 1971-08-24          40.8         54034
#> 4: 52.38536 -2.337805 1971-08-24          40.8         54034
#> 5: 52.47315 -2.197148 1972-08-01          89.9         54063
#> 6: 52.26630 -2.191539 1987-12-10          53.0         54096
#> 7: 52.25542 -2.193836 1960-12-19         184.0         54011
#> 8: 52.25542 -2.193836 1960-12-19         184.0         54011

loadAPI(ID = 'L1207')
#>    parameter period          type
#> 1:     level    900 instantaneous
#> 2:     level  86400           min
#> 3:     level  86400           max

loadAPI(ID = 'L1207',
measure = 'level',
period = 900,
type = 'instantaneous',
datapoints = 'earliest')
#>      dateTime value quality
#> 1: 2003-03-09 0.811    Good

dt <- loadAPI(ID = 'L1207',
measure = 'level',
period = 900,
type = 'instantaneous',
datapoints = 'all')

ggplot2::ggplot(dt, aes(x = dateTime, y = value, colour = factor(quality), group = 1))+
geom_line() +
  xlab('Time') +
    ylab('Stage (mAoD)')
#> Error in aes(x = dateTime, y = value, colour = factor(quality), group = 1): could not find function "aes"