loadWISKI

Load data downloaded from WISKI, in the EA csv, format. Metadata are preserved. Loaded data are in the form of a list, called either flowLoad, rainLoad, or stageLoad. Data are in two parts MetaData and GaugeData.

loadWISKI(
  link,
  skip = 0,
  meta_rows = 15,
  hydro_year = "oct_us_gb",
  cols = NULL
)

Arguments

link

Link to the specified file for import

skip

Stet to zero, denotes the number off rows you wish to skip

meta_rows

Set to 15, determines the amount of rows set as metadata

hydro_year

Set to 'oct_us_gb' (USA and UK), but can also be 'sep_br' (Brazil),'apr_cl' (Chille).

cols

Set to NULL, but can be set as a vector of column names. Note the first 2 must be 'DateTime' and 'Value'

Value

Examples

csvfile <- 'C:/Users/jpayne05/Downloads/Wellingborough.csv'
a <- loadWISKI(csvfile)
#> Importing meta data
#> Importing precipitation data
#>   OpenMP version (_OPENMP)       201511
#>   omp_get_num_procs()            8
#>   R_DATATABLE_NUM_PROCS_PERCENT  unset (default 50)
#>   R_DATATABLE_NUM_THREADS        unset
#>   R_DATATABLE_THROTTLE           unset (default 1024)
#>   omp_get_thread_limit()         2147483647
#>   omp_get_max_threads()          8
#>   OMP_THREAD_LIMIT               unset
#>   OMP_NUM_THREADS                unset
#>   RestoreAfterFork               true
#>   data.table is using 4 threads with throttle==1024. See ?setDTthreads.
#> freadR.c has been passed a filename: C:/Users/jpayne05/Downloads/Wellingborough.csv
#> [01] Check arguments
#>   Using 4 threads (omp_get_max_threads()=8, nth=4)
#>   NAstrings = [<<--->>, <<NA>>]
#>   None of the NAstrings look like numbers.
#>   skip num lines = 15
#>   show progress = 1
#>   0/1 column will be read as integer
#> [02] Opening the file
#>   Opening file C:/Users/jpayne05/Downloads/Wellingborough.csv
#>   File opened, size = 49.15MB (51541360 bytes).
#>   Memory mapped ok
#> [03] Detect and skip BOM
#> [04] Arrange mmap to be \0 terminated
#>   \n has been found in the input and different lines can end with different line endings (e.g. mixed \n and \r\n in one file). This is common and ideal.
#> [05] Skipping initial rows if needed
#>   Skipped to line 16 in the file  Positioned on line 16 starting: <<Time stamp,Value [mm],State of>>
#> [06] Detect separator, quoting rule, and ncolumns
#>   Using supplied sep ','
#>   sep=','  with 5 fields using quote rule 0
#>   Detected 5 columns on line 16. This line is either column names or first data row. Line starts as: <<Time stamp,Value [mm],State of>>
#>   Quote rule picked = 0
#>   fill=true and the most number of columns found is 5
#> [07] Detect column types, good nrow estimate and whether first row is column names
#>   'header' changed by user from 'auto' to true
#>   Number of sampling jump points = 100 because (51540979 bytes from row 1 to eof) / (2 * 4103 jump0size) == 6280
#>   Type codes (jump 000)    : D8D11  Quote rule 0
#>   Type codes (jump 100)    : D8D11  Quote rule 0
#>   =====
#>   Sampled 10048 rows (handled \n inside quoted fields) at 101 jump points
#>   Bytes from first data row on line 17 to the end of last row: 51540935
#>   Line length: mean=41.51 sd=0.93 min=40 max=45
#>   Estimated number of rows: 51540935 / 41.51 = 1241715
#>   Initial alloc = 1365886 rows (1241715 + 9%) using bytes/max(mean-2*sd,min) clamped between [1.1*estn, 2.0*estn]
#>   =====
#> [08] Assign column names
#> [09] Apply user overrides on column types
#>   After 0 type and 0 drop user overrides : D8D11
#> [10] Allocate memory for the datatable
#>   Allocating 5 column slots (5 - 0 dropped) with 1365886 rows
#> [11] Read the data
#>   jumps=[0..52), chunk_size=991171, total_size=51540935
#> Read 1240597 rows x 5 columns from 49.15MB (51541360 bytes) file in 00:01.302 wall clock time
#> [12] Finalizing the datatable
#>   Type counts:
#>          2 : bool8     '1'
#>          1 : float64   '8'
#>          2 : string    'D'
#> =============================
#>    0.002s (  0%) Memory map 0.048GB file
#>    0.065s (  5%) sep=',' ncol=5 and header detection
#>    0.000s (  0%) Column type detection using 10048 sample rows
#>    0.137s ( 11%) Allocation of 1365886 rows x 5 cols (0.041GB) of which 1240597 ( 91%) rows used
#>    1.099s ( 84%) Reading 52 chunks (0 swept) of 0.945MB (each chunk 23857 rows) using 4 threads
#>    +    0.172s ( 13%) Parse to row-major thread buffers (grown 0 times)
#>    +    0.852s ( 65%) Transpose
#>    +    0.074s (  6%) Waiting
#>    0.000s (  0%) Rereading 0 columns due to out-of-sample type exceptions
#>    1.302s        Total
#> Removing blank elements at top
#> Converting dates and times